requirements.txt
pytest == 7.3.1
selenium == 4.9.1
allure-pytest == 2.13.2
webdriver-manager == 3.8.6
Faker~=18.9.0
requests~=2.30.0